+++ src/me/itzsomebody/radon/templates/HeavyStringEncryption.java
@@ -71,10 +71,10 @@ public class HeavyStringEncryption {
                                     broken[1] = ((String) strToDecrypt).toCharArray()[1];
                                     broken[2] = ((String) strToDecrypt).toCharArray()[2];
                                     broken[3] = ((String) strToDecrypt).toCharArray()[3];
-                                    ste = Thread.currentThread().getStackTrace()[one | 255];
+                                    ste = new Throwable().getStackTrace()[one | 255];
                                     throw null;
                                 } catch (Throwable t) {
-                                    ste = Thread.currentThread().getStackTrace()[one];
+                                    ste = new Throwable().getStackTrace()[one - 1];
                                     break label_04;
                                 }
                             }