+++ src/me/itzsomebody/radon/transformers/renamer/Renamer.java
@@ -10,9 +10,7 @@ import org.objectweb.asm.commons.Remapper;
 import org.objectweb.asm.commons.SimpleRemapper;
 import org.objectweb.asm.tree.*;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
@@ -26,6 +24,11 @@ public class Renamer extends AbstractTransformer {
      * Used for class, method and field renamming. Format is oldName -> newName.
      */
     private Map<String, String> mappings = new HashMap<>();
+
+    /**
+     * Used to determine how the classes interact with each other.
+     */
+    private Map<String, ClassTree> hierarchy = new HashMap<>();
 
     /**
      * Indication to look for Bukkit/Bungee main methods.
@@ -45,151 +48,39 @@ public class Renamer extends AbstractTransformer {
     public void obfuscate() {
         logStrings.add(LoggerUtils.stdOut("------------------------------------------------"));
         logStrings.add(LoggerUtils.stdOut("Generating mappings."));
+        createTrees();
         long current = System.currentTimeMillis();
         AtomicInteger counter = new AtomicInteger();
-        classNodes().stream().filter(classNode -> !classExempted(classNode.name)).forEach(classNode -> {
-            classNode.methods.stream().filter(methodNode -> !methodExempted(classNode.name + '.' + methodNode.name + '.' + methodNode.desc))
-                    .filter(methodNode -> !BytecodeUtils.isNativeMethod(methodNode.access)
-                            && !methodNode.name.equals("main") && !methodNode.name.equals("premain")
-                            && !methodNode.name.startsWith("<")
-                            && !mappings.containsKey(classNode.name + '.' + methodNode.name + methodNode.desc)).forEach(methodNode -> {
-                boolean doNotRename = false;
-                if (classNode.superName != null) {
-                    ClassNode superClass = getClassNode(classNode.superName);
-                    outer: {
-                        while (superClass != null && superClass.superName != null) {
-                            for (MethodNode superMethod : superClass.methods) {
-                                if (methodNode.name.equals(superMethod.name)
-                                        && methodNode.desc.equals(superMethod.desc)) {
-                                    if (superClass.libraryNode) {
-                                        doNotRename = true;
-                                        break outer;
-                                    }
-                                }
-                            }
-
-                            if (superClass.interfaces != null) {
-                                for (String className : superClass.interfaces) {
-                                    ClassNode interfaceClass = getClassNode(className);
-                                    for (MethodNode interfaceMethod : interfaceClass.methods) {
-                                        if (methodNode.name.equals(interfaceMethod.name)
-                                                && methodNode.desc.equals(interfaceMethod.desc)) {
-                                            if (interfaceClass.libraryNode) {
-                                                doNotRename = true;
-                                                break outer;
-                                            }
-                                        }
-                                    }
-
-                                /*if (interfaceClass.interfaces != null) {
-                                    for (String subInterface : interfaceClass.interfaces) {
-                                        ClassNode subInterfaceClass = getClassNode(subInterface);
-                                        while (interfaceClass.interfaces != null || !doNotRename) {
-                                            deletethis.println("Starting while loop 2");
-                                            for (MethodNode subInterfaceMethod : subInterfaceClass.methods) {
-                                                if (methodNode.name.equals(subInterfaceClass.name)
-                                                        && methodNode.desc.equals(subInterfaceMethod.desc)) {
-                                                    if (subInterfaceClass.libraryNode) {
-                                                        deletethis.println("The interface class was a lib, don't setting doNotRename");
-                                                        doNotRename = true;
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }*/
-                                }
-                            }
-
-                            superClass = getClassNode(superClass.superName);
-                        }
-                    }
-                }
-
-                outer: {
-                    if (!doNotRename && classNode.interfaces != null) {
-                        for (String className : classNode.interfaces) {
-                            ClassNode interfaceClass = getClassNode(className);
-                            //while (interfaceClass != null && interfaceClass.interfaces != null && !doNotRename) {
-                            if (!interfaceClass.methods.isEmpty()) {
-                                for (MethodNode interfaceMethod : interfaceClass.methods) {
-                                    if (methodNode.name.equals(interfaceMethod.name)
-                                            && methodNode.desc.equals(interfaceMethod.desc)) {
-                                        if (interfaceClass.libraryNode) {
-                                            doNotRename = true;
-                                            break outer;
-                                        }
-                                    }
-                                }
-                            }
-                            //}
-                        }
-                    }
-                }
-
-                if (!doNotRename) {
+        classNodes().forEach(classNode -> {
+            classNode.methods.stream().filter(methodNode -> !BytecodeUtils.isNativeMethod(methodNode.access)
+                    && !methodNode.name.equals("main")
+                    && !methodNode.name.equals("premain")
+                    && !methodNode.name.startsWith("<")
+                    && !methodNode.name.contains("lambda")).forEach(methodNode -> {
+                if (weCanRenameMethod(methodNode)) {
                     String newName = StringUtils.crazyString();
-                    mappings.put(classNode.name + '.' + methodNode.name + methodNode.desc, newName);
-                    ClassNode superClass = getClassNode(classNode.superName);
-                    while (superClass != null && superClass.superName != null) {
-                        for (MethodNode superMethod : superClass.methods) {
-                            if (methodNode.name.equals(superMethod.name)
-                                    && methodNode.desc.equals(superMethod.desc)) {
-                                mappings.put(superClass.name + '.' + methodNode.name + methodNode.desc, newName);
-                            }
-                        }
-
-                        for (String className : superClass.interfaces) {
-                            ClassNode interfaceClass = getClassNode(className);
-                            //while (interfaceClass != null && interfaceClass.interfaces != null) {
-                            for (MethodNode interfaceMethod : interfaceClass.methods) {
-                                if (methodNode.name.equals(interfaceMethod.name)
-                                        && methodNode.desc.equals(interfaceMethod.desc)) {
-                                    mappings.put(interfaceClass.name + '.' + methodNode.name + methodNode.desc, newName);
-                                }
-                                //}
-                            }
-                        }
-
-                        superClass = getClassNode(superClass.superName);
-                    }
-
-                    if (classNode.interfaces != null) {
-                        for (String className : classNode.interfaces) {
-                            ClassNode interfaceClass = getClassNode(className);
-                            //while (interfaceClass != null && interfaceClass.interfaces != null) {
-                            for (MethodNode interfaceMethod : interfaceClass.methods) {
-                                if (methodNode.name.equals(interfaceMethod.name)
-                                        && methodNode.desc.equals(interfaceMethod.desc)) {
-                                    mappings.put(interfaceClass.name + '.' + methodNode.name + methodNode.desc, newName);
-                                }
-                            }
-                            //  }
-                        }
-                    }
+                    renameMethodTree(new ArrayList<>(), methodNode, classNode.name, newName);
                 }
             });
 
-            if (classNode.fields != null) classNode.fields.stream()
-                    .filter(fieldNode -> !fieldExempted(classNode.name + '.' + fieldNode.name)).forEach(fieldNode -> {
-                        mappings.put(classNode.name + '.' + fieldNode.name, StringUtils.crazyString());
-                        counter.incrementAndGet();
-                    });
-
-            if (!BytecodeUtils.isMain(classNode, spigotMode)) {
-                switch (NumberUtils.getRandomInt(3)) {
-                    case 0:
-                        mappings.put(classNode.name, StringUtils.crazyString());
-                        break;
-                    case 1:
-                        mappings.put(classNode.name, StringUtils.crazyString() + '/' + StringUtils.crazyString());
-                        break;
-                    case 2:
-                        mappings.put(classNode.name, StringUtils.crazyString() + '/' + StringUtils.crazyString() + '/' + StringUtils.crazyString());
-                        break;
+            classNode.fields.forEach(fieldNode -> {
+                if (weCanRenameField(fieldNode)) {
+                    String newName = StringUtils.crazyString();
+                    renameFieldTree(new ArrayList<>(), fieldNode, classNode.name, newName);
                 }
+            });
+
+            if (!classExempted(classNode.name)
+                    && !BytecodeUtils.isMain(classNode, spigotMode)) {
+                int packages = NumberUtils.getRandomInt(2) + 1;
+                String newName = "";
+                for (int i = 0; i < packages; i++) {
+                    newName += StringUtils.crazyString() + '/';
+                }
+
+                mappings.put(classNode.name, newName.substring(0, newName.length() - 1));
+                counter.incrementAndGet();
             }
-            counter.incrementAndGet();
         });
         logStrings.add(LoggerUtils.stdOut("Finished generated mappings. [" + String.valueOf(System.currentTimeMillis() - current) + "ms]"));
         logStrings.add(LoggerUtils.stdOut("Applying mappings."));
@@ -198,75 +89,6 @@ public class Renamer extends AbstractTransformer {
         // Apply mapping
         Remapper simpleRemapper = new SimpleRemapper(mappings);
         for (ClassNode classNode : new ArrayList<>(classNodes())) {
-            // Now we have to fix inheritance issues which ASM doesn't freaking fix
-            for (MethodNode mn : classNode.methods) {
-                if (classNode.superName == null
-                        && classNode.interfaces == null) {
-                    break; // No inheritance whatsoever
-                }
-                for (AbstractInsnNode insn : mn.instructions.toArray()) {
-                    if (insn instanceof MethodInsnNode) {
-                        if (insn.getOpcode() == INVOKEVIRTUAL
-                                || insn.getOpcode() == INVOKEINTERFACE) {
-                            MethodInsnNode min = (MethodInsnNode) insn;
-                            if (min.owner.equals(classNode.name)
-                                    && !BytecodeUtils.containsMethod(min.name, min.desc, classNode)) {
-                                ClassNode superClass = getClassNode(classNode.superName);
-                                if (!superClass.libraryNode) { // Lib node = don't touch.
-                                    if (BytecodeUtils.containsMethod(min.name, min.desc, superClass)) {
-                                        String key = superClass.name + '.' + min.name + min.desc;
-                                        if (mappings.containsKey(key)) {
-                                            mappings.put(classNode.name + '.' + min.name + min.desc, mappings.get(key));
-                                        }
-                                    }
-                                }
-
-                                for (String interfaceName : classNode.interfaces) {
-                                    ClassNode interfaceClass = getClassNode(interfaceName);
-                                    if (!interfaceClass.libraryNode) { // Lib node = don't touch.
-                                        if (BytecodeUtils.containsMethod(min.name, min.desc, interfaceClass)) {
-                                            String key = interfaceClass.name + '.' + min.name + min.desc;
-                                            if (mappings.containsKey(key)) {
-                                                mappings.put(classNode.name + '.' + min.name + min.desc, mappings.get(key));
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    } else if (insn instanceof FieldInsnNode) {
-                        if (insn.getOpcode() == GETFIELD
-                                || insn.getOpcode() == PUTFIELD) {
-                            FieldInsnNode fin = (FieldInsnNode) insn;
-                            if (fin.owner.equals(classNode.name)
-                                    && !BytecodeUtils.containsField(fin.name, fin.desc, classNode)) {
-                                ClassNode superClass = getClassNode(classNode.superName);
-                                if (!superClass.libraryNode) { // Lib node = don't touch.
-                                    if (BytecodeUtils.containsField(fin.name, fin.desc, superClass)) {
-                                        String key = superClass.name + '.' + fin.name;
-                                        if (mappings.containsKey(key)) {
-                                            mappings.put(classNode.name + '.' + fin.name, mappings.get(key));
-                                        }
-                                    }
-                                }
-
-                                for (String interfaceName : classNode.interfaces) {
-                                    ClassNode interfaceClass = getClassNode(interfaceName);
-                                    if (!interfaceClass.libraryNode) { // Lib node = don't touch.
-                                        if (BytecodeUtils.containsMethod(fin.name, fin.desc, interfaceClass)) {
-                                            String key = interfaceClass.name + '.' + fin.name;
-                                            if (mappings.containsKey(key)) {
-                                                mappings.put(classNode.name + '.' + fin.name, mappings.get(key));
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
             ClassNode copy = new ClassNode();
             classNode.accept(new ClassRemapper(copy, simpleRemapper));
             copy.access = BytecodeUtils.accessFixer(copy.access);
@@ -285,7 +107,175 @@ public class Renamer extends AbstractTransformer {
             getClassMap().put(copy.name, copy);
         }
 
-        logStrings.add(LoggerUtils.stdOut("Renamed " + counter + " members."));
+        logStrings.add(LoggerUtils.stdOut("Mapped " + counter + " members."));
         logStrings.add(LoggerUtils.stdOut("Finished applying mappings. [" + String.valueOf(System.currentTimeMillis() - current) + "ms]"));
+    }
+
+    /**
+     * Creates {@link ClassTree}s needed for renaming.
+     */
+    private void createTrees() {
+        logStrings.add(LoggerUtils.stdOut("Creating class hierarchy."));
+        long current = System.currentTimeMillis();
+        getClassPathMap().values().forEach(classNode -> {
+            classNode.methods.forEach(methodNode -> {
+                methodNode.owner = classNode.name;
+            });
+            classNode.fields.forEach(fieldNode -> {
+                fieldNode.owner = classNode.name;
+            });
+            ClassTree classTree = new ClassTree(classNode.name, classNode.libraryNode);
+            classTree.parentClasses.add(classNode.superName);
+            classTree.parentClasses.addAll(classNode.interfaces);
+            classNodes().stream().filter(node -> node.superName.equals(classNode.name)
+                    || node.interfaces.contains(classNode.name)).forEach(node -> {
+                classTree.subClasses.add(node.name);
+            });
+
+            classTree.methods.addAll(classNode.methods);
+            classTree.fields.addAll(classNode.fields);
+            hierarchy.put(classNode.name, classTree);
+        });
+        logStrings.add(LoggerUtils.stdOut("Finished creating class hierarchy. [" + tookThisLong(current) + "ms]"));
+    }
+
+    /**
+     * Ultimately determines if we can rename a method without running into errors.
+     *
+     * @param methodNode {@link MethodNode} we want to check if we can rename without causing the JVM to spit lots of errors.
+     * @return true if we can rename a method without running into errors.
+     */
+    private boolean weCanRenameMethod(MethodNode methodNode) {
+        if (methodExempted(methodNode.owner + '.' + methodNode.name + methodNode.desc)) {
+            return false;
+        }
+        for (ClassTree ct : hierarchy.values()) {
+            if (ct.subClasses.contains(methodNode.owner)
+                    && isLibInheritedMN(methodNode, methodNode.owner)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /**
+     * Attempts to determine if the method we input is inherited from a library class.
+     *
+     * @param methodNode {@link MethodNode} we want to check if we can rename without causing the JVM to spit lots of errors.
+     * @param className  the name of the class we want to check
+     * @return true if the method we input is inherited from a library class.
+     */
+    private boolean isLibInheritedMN(MethodNode methodNode, String className) {
+        ClassTree ct = hierarchy.get(className);
+        if (ct != null) {
+            if (!methodNode.owner.equals(className)) {
+                for (MethodNode mn : ct.methods) {
+                    if (mn.name.equals(methodNode.name) && mn.desc.equals(methodNode.desc)) {
+                        if (ct.libraryNode) {
+                            return true;
+                        }
+                    }
+                }
+            }
+            for (String parentClass : ct.parentClasses) {
+                if (isLibInheritedMN(methodNode, parentClass)) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Renames the methods in an inheritance tree to prevent inheritance errors.
+     *
+     * @param visited    a list of {@link ClassTree}s which contain the {@link MethodNode}s we already renamed.
+     * @param methodNode the method information.
+     * @param className  the class we are currently browsing through.
+     * @param newName    the new name of the method.
+     */
+    private void renameMethodTree(List<ClassTree> visited, MethodNode methodNode, String className, String newName) {
+        ClassTree ct = hierarchy.get(className);
+        if (!ct.libraryNode && !visited.contains(ct)) {
+            mappings.put(className + '.' + methodNode.name + methodNode.desc, newName);
+            visited.add(ct);
+            for (String parentClass : ct.parentClasses) {
+                renameMethodTree(visited, methodNode, parentClass, newName);
+            }
+            for (String subClass : ct.subClasses) {
+                renameMethodTree(visited, methodNode, subClass, newName);
+            }
+        }
+    }
+
+    /**
+     * Ultimately determines if we can rename a field without running into errors.
+     *
+     * @param fieldNode {@link FieldNode} we want to check if we can rename without causing the JVM to spit lots of errors.
+     * @return true if we can rename a field without running into errors.
+     */
+    private boolean weCanRenameField(FieldNode fieldNode) {
+        if (methodExempted(fieldNode.owner + '.' + fieldNode.name)) {
+            return false;
+        }
+        for (ClassTree ct : hierarchy.values()) {
+            if (ct.subClasses.contains(fieldNode.owner)
+                    && isLibInheritedFN(fieldNode, fieldNode.owner)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /**
+     * Attempts to determine if the method we input is inherited from a library class.
+     *
+     * @param fieldNode {@link FieldNode} we want to check if we can rename without causing the JVM to spit lots of errors.
+     * @param className  the name of the class we want to check
+     * @return true if the method we input is inherited from a library class.
+     */
+    private boolean isLibInheritedFN(FieldNode fieldNode, String className) {
+        ClassTree ct = hierarchy.get(className);
+        if (ct != null) {
+            if (!fieldNode.owner.equals(className)) {
+                for (MethodNode mn : ct.methods) {
+                    if (mn.name.equals(fieldNode.name) && mn.desc.equals(fieldNode.desc)) {
+                        if (ct.libraryNode) {
+                            return true;
+                        }
+                    }
+                }
+            }
+            for (String parentClass : ct.parentClasses) {
+                if (isLibInheritedFN(fieldNode, parentClass)) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Renames the methods in an inheritance tree to prevent inheritance errors.
+     *
+     * @param visited    a list of {@link ClassTree}s which contain the {@link MethodNode}s we already renamed.
+     * @param fieldNode the method information.
+     * @param className  the class we are currently browsing through.
+     * @param newName    the new name of the method.
+     */
+    private void renameFieldTree(List<ClassTree> visited, FieldNode fieldNode, String className, String newName) {
+        ClassTree ct = hierarchy.get(className);
+        if (!ct.libraryNode && !visited.contains(ct)) {
+            mappings.put(className + '.' + fieldNode.name, newName);
+            visited.add(ct);
+            for (String parentClass : ct.parentClasses) {
+                renameFieldTree(visited, fieldNode, parentClass, newName);
+            }
+            for (String subClass : ct.subClasses) {
+                renameFieldTree(visited, fieldNode, subClass, newName);
+            }
+        }
     }
 }